requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-migrate==4.0.5
werkzeug==3.0.1
pillow==10.2.0
python-dotenv==1.0.1
bcrypt==4.1.2
timm==0.9.12